Restructuring and cleaning up, add windows support
