+++ rviz_python_bindings/CMakeLists.txt
@@ -1,8 +1,7 @@
-cmake_minimum_required(VERSION 3.8)
+cmake_minimum_required(VERSION 3.17)
 project(rviz_python_bindings)
 
-set(bindings_library "rviz_shiboken")
-
+#find_package (Python3 COMPONENTS Interpreter Development)
 find_package(ament_cmake REQUIRED)
 find_package(ament_cmake_python REQUIRED)
 find_package(Shiboken2 REQUIRED)
@@ -11,6 +10,27 @@ find_package(rviz_common REQUIRED)
 find_package(rviz_ogre_vendor REQUIRED)
 find_package(pluginlib REQUIRED)
 
+
+# name of the library to be generated by shiboken
+set(bindings_library "rviz_shiboken")
+# path to the include wrapper used by shiboken for generation
+set(include_wrapper "${CMAKE_CURRENT_SOURCE_DIR}/conf/global.hpp")
+# path to the typesystem file used by shiboken for generation
+set(typesystem "${CMAKE_CURRENT_SOURCE_DIR}/conf/typesystem.xml")
+# sources generated by shiboken depends on typesystem
+set(generated_sources
+    ${CMAKE_CURRENT_BINARY_DIR}/${bindings_library}/rviz_shiboken_module_wrapper.cpp
+    ${CMAKE_CURRENT_BINARY_DIR}/${bindings_library}/rviz_common_wrapper.cpp
+    ${CMAKE_CURRENT_BINARY_DIR}/${bindings_library}/rviz_common_visualizerframepy_wrapper.cpp
+    ${CMAKE_CURRENT_BINARY_DIR}/${bindings_library}/rviz_common_yamlconfigreader_wrapper.cpp
+    ${CMAKE_CURRENT_BINARY_DIR}/${bindings_library}/rviz_common_yamlconfigwriter_wrapper.cpp
+    ${CMAKE_CURRENT_BINARY_DIR}/${bindings_library}/rviz_common_config_wrapper.cpp
+    ${CMAKE_CURRENT_BINARY_DIR}/${bindings_library}/rviz_common_visualizationframe_wrapper.cpp
+    ${CMAKE_CURRENT_BINARY_DIR}/${bindings_library}/rviz_common_visualizationmanager_wrapper.cpp
+    ${CMAKE_CURRENT_BINARY_DIR}/${bindings_library}/rviz_common_displaygroup_wrapper.cpp
+    ${CMAKE_CURRENT_BINARY_DIR}/${bindings_library}/rviz_common_display_wrapper.cpp
+    ${CMAKE_CURRENT_BINARY_DIR}/${bindings_library}/rviz_common_viewmanager_wrapper.cpp
+)
 
 add_library(rviz_python SHARED
   src/visualizer_frame_py.cpp
@@ -32,7 +52,29 @@ install(TARGETS rviz_python
   RUNTIME DESTINATION bin
 )
 
-message(STATUS "LIBRARY OUTPUT DIR ${CMAKE_INSTALL_PREFIX}/${PYTHON_INSTALL_DIR}/rviz")
+# macro to extract information from Windows systems using pyside_config.py
+macro(pyside_config option output_var)
+    if(${ARGC} GREATER 2)
+        set(is_list ${ARGV2})
+    else()
+        set(is_list "")
+    endif()
+
+    execute_process(
+      COMMAND ${Python3_EXECUTABLE} "${CMAKE_SOURCE_DIR}/pyside_config.py"
+              ${option}
+      OUTPUT_VARIABLE ${output_var}
+      OUTPUT_STRIP_TRAILING_WHITESPACE)
+
+    if ("${${output_var}}" STREQUAL "")
+        message(FATAL_ERROR "Error: Calling pyside_config.py ${option} returned no output.")
+    endif()
+    if(is_list)
+        string (REPLACE " " ";" ${output_var} "${${output_var}}")
+    endif()
+endmacro()
+
+# Find Shiboken
 if(Shiboken2_FOUND)
   message(STATUS "Found Shiboken2 version ${Shiboken2_VERSION}")
   if(NOT ${Shiboken2_VERSION} VERSION_LESS "5.13")
@@ -40,11 +82,23 @@ if(Shiboken2_FOUND)
     get_property(SHIBOKEN_LIBRARY TARGET Shiboken2::libshiboken PROPERTY LOCATION)
     set(SHIBOKEN_BINARY Shiboken2::shiboken2)
   endif()
-  message(STATUS "Using SHIBOKEN_INCLUDE_DIR: ${SHIBOKEN_INCLUDE_DIR}")
-  message(STATUS "Using SHIBOKEN_LIBRARY: ${SHIBOKEN_LIBRARY}")
-  message(STATUS "Using SHIBOKEN_BINARY: ${SHIBOKEN_BINARY}")
+else()
+  if(WIN32)
+    pyside_config(--shiboken-generator-path shiboken_generator_path)
+    pyside_config(--shiboken-generator-include-path shiboken_include_dir 1)
+    pyside_config(--shiboken-module-shared-libraries-cmake shiboken_shared_libraries 0)
+
+    set(SHIBOKEN_BINARY "${shiboken_generator_path}/shiboken2${CMAKE_EXECUTABLE_SUFFIX}")
+    set(SHIBOKEN_LIBRARY ${shiboken_shared_libraries})
+    set(SHIBOKEN_INCLUDE_DIR ${shiboken_include_dir})
+  endif()
 endif()
 
+message(STATUS "Using SHIBOKEN_INCLUDE_DIR: ${SHIBOKEN_INCLUDE_DIR}")
+message(STATUS "Using SHIBOKEN_LIBRARY: ${SHIBOKEN_LIBRARY}")
+message(STATUS "Using SHIBOKEN_BINARY: ${SHIBOKEN_BINARY}")
+
+# Find Pyside
 find_package(PySide2 QUIET)
 if(PySide2_FOUND)
   message(STATUS "Found PySide2 version ${PySide2_VERSION}")
@@ -52,26 +106,23 @@ if(PySide2_FOUND)
     get_property(PYSIDE_INCLUDE_DIR TARGET PySide2::pyside2 PROPERTY INTERFACE_INCLUDE_DIRECTORIES)
     get_property(PYSIDE_LIBRARY TARGET PySide2::pyside2 PROPERTY LOCATION)
   endif()
-  message(STATUS "Using PYSIDE_INCLUDE_DIR: ${PYSIDE_INCLUDE_DIR}")
-  message(STATUS "Using PYSIDE_LIBRARY: ${PYSIDE_LIBRARY}")
-  message(STATUS "Using PYSIDE_TYPESYSTEMS: ${PYSIDE_TYPESYSTEMS}")
+else ()
+  if(WIN32)
+    pyside_config(--pyside-include-path pyside_include_dir 1)
+    pyside_config(--pyside-shared-libraries-cmake pyside_shared_libraries 0)
+    pyside_config(--pyside-path pyside_path)
+    set(PYSIDE_TYPESYSTEMS "${pyside_path}/typesystems")
+    set(PYSIDE_INCLUDE_DIR ${pyside_include_dir})
+    set(PYSIDE_LIBRARY ${pyside_shared_libraries})
+  endif()
 endif()
 
-set(generated_sources
-    ${CMAKE_CURRENT_BINARY_DIR}/${bindings_library}/rviz_shiboken_module_wrapper.cpp
-    ${CMAKE_CURRENT_BINARY_DIR}/${bindings_library}/rviz_common_wrapper.cpp
-    ${CMAKE_CURRENT_BINARY_DIR}/${bindings_library}/rviz_common_visualizerframepy_wrapper.cpp
-    ${CMAKE_CURRENT_BINARY_DIR}/${bindings_library}/rviz_common_yamlconfigreader_wrapper.cpp
-    ${CMAKE_CURRENT_BINARY_DIR}/${bindings_library}/rviz_common_yamlconfigwriter_wrapper.cpp
-    ${CMAKE_CURRENT_BINARY_DIR}/${bindings_library}/rviz_common_config_wrapper.cpp
-    ${CMAKE_CURRENT_BINARY_DIR}/${bindings_library}/rviz_common_visualizationframe_wrapper.cpp
-    ${CMAKE_CURRENT_BINARY_DIR}/${bindings_library}/rviz_common_visualizationmanager_wrapper.cpp
-    ${CMAKE_CURRENT_BINARY_DIR}/${bindings_library}/rviz_common_displaygroup_wrapper.cpp
-    ${CMAKE_CURRENT_BINARY_DIR}/${bindings_library}/rviz_common_display_wrapper.cpp
-    )
+message(STATUS "Using PYSIDE_INCLUDE_DIR: ${PYSIDE_INCLUDE_DIR}")
+message(STATUS "Using PYSIDE_LIBRARY: ${PYSIDE_LIBRARY}")
+message(STATUS "Using PYSIDE_TYPESYSTEMS: ${PYSIDE_TYPESYSTEMS}")
+
 
 list(APPEND SHIBOKEN_HELPER_INCLUDE_DIRS
-    ${QT_INCLUDE_DIR}
     ${PYSIDE_INCLUDE_DIR}
     ${PYTHON_INCLUDE_DIR}
     ${SHIBOKEN_INCLUDE_DIR}
@@ -106,9 +157,9 @@ set(shiboken_options --generator-set=shiboken
 
 add_custom_command(OUTPUT ${generated_sources}
     COMMAND ${SHIBOKEN_BINARY}
-    ${shiboken_options} ${CMAKE_CURRENT_SOURCE_DIR}/global.hpp ${CMAKE_CURRENT_SOURCE_DIR}/typesystem.xml
+    ${shiboken_options} ${include_wrapper} ${typesystem}
     DEPENDS ${generated_sources_dependencies}
-    IMPLICIT_DEPENDS CXX global.hpp
+    IMPLICIT_DEPENDS CXX ${include_wrapper}
     WORKING_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR}
     COMMENT "Running generator for ${typesystem_file}.")
 
@@ -141,8 +192,7 @@ target_include_directories(${bindings_library} PRIVATE
   )
 
 target_link_libraries(${bindings_library} PRIVATE
-  ${QT_LIBRARIES}
-  ${PYTHON_LIBRARY}
+  ${Python3_LIBRARIES}
   ${SHIBOKEN_LIBRARY}
   ${PYSIDE_LIBRARY}
   Qt5::Core
@@ -154,6 +204,15 @@ target_link_libraries(${bindings_library} PRIVATE
 
 set_target_properties(${bindings_library} PROPERTIES
 PREFIX "")
+set_property(TARGET ${bindings_library} PROPERTY OUTPUT_NAME
+             "${bindings_library}${Python3_SOABI}")
+if(WIN32)
+    if("${CMAKE_BUILD_TYPE}" STREQUAL "Debug")
+        set_property(TARGET ${bindings_library} PROPERTY SUFFIX "_d.pyd")
+    else()
+        set_property(TARGET ${bindings_library} PROPERTY SUFFIX ".pyd")
+    endif()
+endif()
 
 
 install(TARGETS ${bindings_library}
