+++ rviz_python_bindings/README.md
@@ -1,0 +1,12 @@
+# RVIZ Pyhon Bindings
+
+
+## Windows Dependencies
+On windows you will need to install the following dependencies manually:
+
+```
+pip install \
+    --index-url=https://download.qt.io/official_releases/QtForPython/ \
+    --trusted-host download.qt.io \
+    shiboken2 pyside2 shiboken2_generator
+```
